changed order of loops through lambda and random_states + other changes
This version runs!!

Changed the order of the loops through lambda and random_states, in order to generate new boots for a given lambda. In the previous version, the fit for the bootstrap subset was done for the best lambda (and this changed during every loop).

More changes in this version:
1) Calculate z_pred as the average of the predicted z's for each boot, and its mean (over all boots).

2) Changed the calculation of MSE ase the sum of each MSE* in a boot divided by the number of boots.

3) Removed random_state forme everywher except from the bootstrap loop and algorithms.py.

4) Mooved most of the functions in model_comparison.py to a new script called functions.py